% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punct.freq.R
\name{punct.freq}
\alias{punct.freq}
\title{punct.freq}
\usage{
punct.freq(text)
}
\arguments{
\item{text}{This should be a text (.txt) file.}
}
\value{
a list of punctuation frequencies
}
\description{
Provides a list of all of the punctuation in the .txt file.
}
\examples{
punct.freq(text)
}
