% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punct.hist.R
\name{punct.hist}
\alias{punct.hist}
\title{punct.hist}
\usage{
punct.hist(p.freq, n = 20, print.val = FALSE)
}
\arguments{
\item{p.freq}{This should be the list of punctuations frequencies (frequency function).}

\item{n}{This indicates the top x most frequently occuring punctuations that will be graphed}

\item{print.val}{This indicates whether a list of the top n punctuations should be printed out}
}
\value{
a histogram of the top n frequently occuring punctuations in a text (can also return a list of the top n frequently occuring words)
}
\description{
Given a list of punctuations and frequencies and an integer n, plot a histogram of top n punctuations.
}
\examples{
punct.hist(p.freq, 50, TRUE,)
}
