% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sent.avg.R
\name{sent.avg}
\alias{sent.avg}
\title{sent.avg}
\usage{
sent.avg(text)
}
\arguments{
\item{text}{This should be a text (.txt) file.}
}
\value{
a double indicating the average sentence length used in the text
}
\description{
Provides an average sentence length of the words in the .txt file.
}
\examples{
sent.avg(text)
}
