% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.avg.R
\name{word.avg}
\alias{word.avg}
\title{word.avg}
\usage{
word.avg(text)
}
\arguments{
\item{text}{This should be a text (.txt) file.}
}
\value{
a double indicating the average word length used in the text
}
\description{
Provides an average word length of the words in the .txt file.
}
\examples{
word.avg(text)
}
