% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.filter.R
\name{word.filter}
\alias{word.filter}
\title{word.filter}
\usage{
word.filter(w.freq, include.sw = TRUE)
}
\arguments{
\item{w.freq}{This should be the list of word frequencies (word.freq function).}

\item{include.sw}{This is a boolean indicating whether stopwords should be included: TRUE=stopwords removed, FALSE= stopwords only}
}
\value{
a list of the approriate words and their frequencies
}
\description{
Provides a list of all of the words in the .txt file.
}
\examples{
word.freq(w.freq, include.sw=TRUE)
}
