% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.freq.R
\name{word.freq}
\alias{word.freq}
\title{word.freq}
\usage{
word.freq(text)
}
\arguments{
\item{text}{This should be a text (.txt) file.}
}
\value{
a list of word frequencies
}
\description{
Provides a list of all of the words in the .txt file.
}
\examples{
word.freq(text)
}
