% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.hist.R
\name{word.hist}
\alias{word.hist}
\title{word.hist}
\usage{
word.hist(w.freq, n = 20, print.val = FALSE)
}
\arguments{
\item{w.freq}{This should be the list of word frequencies (3 frequency functions).}

\item{n}{This indicates the top x most frequently occuring words that will be graphed}

\item{print.val}{This indicates whether a list of the top n words should be printed out}
}
\value{
a histogram of the top n frequently occuring words in a text (can also return a list of the top n frequently occuring words)
}
\description{
Given a list of words and frequencies, an integer n, and a boolean indicating whether to print the n values, plot a histogram of top n words.
}
\examples{
word.hist(w.freq, 50, TRUE,)
}
